Link is redirecting to google page problem with CORS
