+++ package.json
@@ -14,6 +14,7 @@
     "connect-assets": "^5.0.1",
     "cookie-parser": "^1.3.3",
     "cookie-session": "^1.2.0",
+    "cors": "^2.8.4",
     "errorhandler": "^1.2.2",
     "express": "^4.9.8",
     "express-flash": "^0.0.2",
